// TimeWarden.Client/timewarden-client/src/api/generated/models/itemOfWorkVM.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * TimeWarden.API | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ProjectVM } from './projectVM';

export interface ItemOfWorkVM {
  id?: string;
  /** @nullable */
  invoiceId?: string | null;
  /** @nullable */
  projectId?: string | null;
  description?: string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  hourlyRate?: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$ */
  hoursOfWork?: number | string;
  dateOfWork?: string;
  project?: null | ProjectVM;
}